Add files via upload
This is my code after finishing the data visualization section of "Eric Matthes Python Crash Course" (I highly recommend doing it. you can get the free pdf online). This code shows examples of some of the things you can do using matplotlib and plotly.

Here is one of the outputs of a function in my code which plots the most starred GitHub repositories that where coded in python: